+++ app/vmui/packages/vmui/src/components/Main/CodeExample/CodeExample.tsx
@@ -4,13 +4,16 @@ import { useState } from "react";
 import Tooltip from "../Tooltip/Tooltip";
 import Button from "../Button/Button";
 import { CopyIcon } from "../Icons";
+import useCopyToClipboard from "../../../hooks/useCopyToClipboard";
 
 enum CopyState { copy = "Copy", copied = "Copied" }
 
 const CodeExample: FC<{code: string}> = ({ code }) => {
+  const copyToClipboard = useCopyToClipboard();
+
   const [tooltip, setTooltip] = useState(CopyState.copy);
-  const handlerCopy = () => {
-    navigator.clipboard.writeText(code);
+  const handlerCopy = async () => {
+    await copyToClipboard(code);
     setTooltip(CopyState.copied);
   };